Move BookTitle to dto.booktitle
